Make type violations more explicit for XML
We generated type violations as unexpected lists for non-lists, and
strings for lists, respectively. This works well in JSON, but not is not
so clear in XML. The lists in XML are serialized as a sequence of XML
elements, which resembles also the serialization of the properties.
Hence, when a list is passed in insteead of an atomic value, the first
element in the list is interpreted as an invalid (additional,
unexpected) property rather than list in XML.

With this change, we explicitly put a string where anything
non-primitive is expected, and a reference where primitive values are
expected. This way we make the type violations more explicit which in
turn makes the debugging in the SDKs easier, since more meaningful
errors are reported.