Make type violations explicit for enumerations (#13)
In #12, we made type violations explicit for non-primitives. However,
for enumerations, the type violations are now misinterpreted as invalid
enumeration literals, as enumerations in JSON and XML are represented as
strings.

In this patch, we generate type violations for enumerations as
unexpected instances to avoid confusion due to uninformative messages
reported from the SDKs.