Change generate scripts to use path arguments
We change the generate scripts so that the model path as well as the
target path of their artefacts need to be specified. This is necessary,
so that the operators can run the scripts independent of their virtual
environment.

For example, when we generate test data for a provisional meta-model, we
want to quickly see the results, without messing up the virtual
environment for this application.