Change generate scripts to use path arguments (#32)
We change the generate scripts so that the model path as well as
the target path of their artifacts need to be specified. This is
necessary, so that the operators can run the scripts independent of
their virtual environment. Previously, we relied on aas-core-meta and
hard-coded the meta-model module, *i.e.*, `aas_core_meta.v3`. Now you
can generate the test data for an arbitrary meta-model, which does not
have to correspond to `aas_core_meta.v3`.

For example, when we generate test data for a provisional meta-model,
we want to quickly see the results, without messing up the virtual
environment for this application.