Split negative examples (#36)
We split the negative examples into unserializable directory and invalid
directory, so that it is easier for the downstream users to test against
them. For example, the SDK developers can test their SDKs directly on
one of the two directories, and check if the outcomes match with the
directory.

In addition, this split also allows us to verify the generated data
against the Python SDK included in this repository. This way we can
immediately reveal many bugs either in our generation procedure or in
the Python SDK.

So far, we tested this in the unit tests tightly coupled to
the meta-model given as a module (and expected to be installed in
the virtual environment). With this change, we can now also test our
logic against an arbitrary V3.0 meta-model supplied as a path
(`--model_path`). For example, this is practical if we want to test
against V3.0.1, which is a patch version of the meta-model and as such
will not live in a separate path in aas-core-meta, but will overwrite
`v3.py`.